__app__/requirements.txt
azure-common==1.1.25
azure-core==1.7.0
azure-nspkg==3.0.2
azure-storage-blob==12.3.2
azure-storage-common==2.1.0
azure-storage-nspkg==3.1.0
azure-functions==1.3.0
google-cloud-bigquery==1.25.0
pyyaml==5.3
pyarrow == 0.17.1
pandas == 1.0.3
requests==2.22.0
click==7.1.2
six==1.13.0
XlsxWriter==1.2.3